python-agent/requirements.txt
argparse
langchain
sqlalchemy
openai
pyyaml
psycopg2